+++ src/main/java/io/asfjava/ui/schema/decorator/PasswordSchemaDecorator.java
@@ -1,17 +1,16 @@
-package io.asf.java.ui.schema.decorator;
+package io.asfjava.ui.schema.decorator;
 
 import com.fasterxml.jackson.databind.BeanProperty;
 import com.fasterxml.jackson.module.jsonSchema.JsonSchema;
 import com.fasterxml.jackson.module.jsonSchema.types.StringSchema;
 
 import io.asfjava.ui.core.form.Password;
-import io.asfjava.ui.core.form.RadioBox;
 
-public class RadioBoxSchemaDecorator implements SchemaDecorator {
+public class PasswordSchemaDecorator implements SchemaDecorator {
 
 	@Override
 	public void customizeSchema(BeanProperty property, JsonSchema jsonschema) {
-		RadioBox annotation = property.getAnnotation(RadioBox.class);
+		Password annotation = property.getAnnotation(Password.class);
 		if (annotation != null && annotation.title() != null) {
 			((StringSchema) jsonschema).setTitle(annotation.title());
 		}
@@ -19,7 +18,7 @@ public class RadioBoxSchemaDecorator implements SchemaDecorator {
 
 	@Override
 	public String getAnnotation() {
-		return RadioBox.class.getName();
+		return Password.class.getName();
 	}
 
 }
