+++ src/main/java/io/asfjava/ui/schema/decorator/RadioBoxSchemaDecorator.java
@@ -1,16 +1,17 @@
-package io.asf.java.ui.schema.decorator;
+package io.asfjava.ui.schema.decorator;
 
 import com.fasterxml.jackson.databind.BeanProperty;
 import com.fasterxml.jackson.module.jsonSchema.JsonSchema;
 import com.fasterxml.jackson.module.jsonSchema.types.StringSchema;
 
 import io.asfjava.ui.core.form.Password;
+import io.asfjava.ui.core.form.RadioBox;
 
-public class PasswordSchemaDecorator implements SchemaDecorator {
+public class RadioBoxSchemaDecorator implements SchemaDecorator {
 
 	@Override
 	public void customizeSchema(BeanProperty property, JsonSchema jsonschema) {
-		Password annotation = property.getAnnotation(Password.class);
+		RadioBox annotation = property.getAnnotation(RadioBox.class);
 		if (annotation != null && annotation.title() != null) {
 			((StringSchema) jsonschema).setTitle(annotation.title());
 		}
@@ -18,7 +19,7 @@ public class PasswordSchemaDecorator implements SchemaDecorator {
 
 	@Override
 	public String getAnnotation() {
-		return Password.class.getName();
+		return RadioBox.class.getName();
 	}
 
 }
