+++ src/main/java/io/asfjava/ui/schema/decorator/TextAreaSchemaDecorator.java
@@ -1,16 +1,16 @@
-package io.asf.java.ui.schema.decorator;
+package io.asfjava.ui.schema.decorator;
 
 import com.fasterxml.jackson.databind.BeanProperty;
 import com.fasterxml.jackson.module.jsonSchema.JsonSchema;
 import com.fasterxml.jackson.module.jsonSchema.types.StringSchema;
 
-import io.asfjava.ui.core.form.ComboBox;
+import io.asfjava.ui.core.form.TextArea;
 
-public class ComboBoxSchemaDecorator implements SchemaDecorator {
+public class TextAreaSchemaDecorator implements SchemaDecorator {
 
 	@Override
 	public void customizeSchema(BeanProperty property, JsonSchema jsonschema) {
-		ComboBox annotation = property.getAnnotation(ComboBox.class);
+		TextArea annotation = property.getAnnotation(TextArea.class);
 		if (annotation != null && annotation.title() != null) {
 			((StringSchema) jsonschema).setTitle(annotation.title());
 		}
@@ -18,7 +18,7 @@ public class ComboBoxSchemaDecorator implements SchemaDecorator {
 
 	@Override
 	public String getAnnotation() {
-		return ComboBox.class.getName();
+		return TextArea.class.getName();
 	}
 
 }
